requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
Pillow>=8.0.0
PyYAML>=6.0
tqdm>=4.62.0
tensorboard>=2.8.0
matplotlib>=3.5.0
